feat(terms): make term.yaml more conformant
Organized term schema more naturally; conforms better to
GDC-JSONschema. _term_ is a real GDC node in this version.

Made "synonyms" an array of foreign key links with a GDC-JSONschema
specification.

Added "values" array to the termdef as a place to hold the enumerated
values of terms that are used as keys elsewhere.