feat(terms): make a schema-valued 'values' field
To address the issue of recording acceptable values, when a term
represents a key, added an optional "values" field to schema. The
value of this field _in the instance_ is a valid JSON schema, which
might be used to validate values when the term is encountered in
the wild. See term_template.yaml for examples.