s/28s_16s_ribosomal_rna_ratio/ribosomal_rna_28s_16s_ratio/g
This change is required because we are injecting property names into
certain places in the GDC.  Specifically, this breaks creation of fields
for Graphene (for Submission GraphQL).

I've also updated the metaschema to validate against a sane property
naming scheme to prevent this in the future and the tests will fail
should a new property that does not match the regex
`^[_a-zA-Z][_a-zA-Z0-9]*$` be added.